refactor: run new stylelint script
