+++ src/app/notes/[id]/components/NoteDetail/index.styled.ts
@@ -2,11 +2,12 @@ import styled from 'styled-components';
 
 export const Containter = styled.section`
   display: flex;
-  justify-content: center;
   flex-direction: column;
-  margin: 2rem auto;
+  justify-content: center;
+
   width: 100%;
   max-width: 75rem;
+  margin: 2rem auto;
 
   h6 {
     font-weight: 400;
@@ -16,14 +17,14 @@ export const Containter = styled.section`
 export const ContentContainter = styled.div`
   display: flex;
   flex-direction: column;
-  align-items: center;
   gap: 2rem;
+  align-items: center;
 
-  margin: 1rem auto;
-  padding: 2rem;
   width: 100%;
   max-width: 65rem;
   min-height: 50vh;
+  margin: 1rem auto;
+  padding: 2rem;
 
   background-color: ${({ theme }) => theme.colors.tertiary};
   border-radius: 1rem;
@@ -35,12 +36,12 @@ export const ContentContainter = styled.div`
 
 export const BottomContainer = styled.div`
   display: flex;
-  justify-content: space-between;
   align-items: center;
+  justify-content: space-between;
 
-  margin: 0 auto;
   width: 100%;
   max-width: 65rem;
+  margin: 0 auto;
   padding: 0 2rem;
 
   svg {
