+++ src/app/notes/components/Note/index.styled.ts
@@ -3,26 +3,29 @@ import styled from 'styled-components';
 
 export const CardContainer = styled.li`
   display: flex;
-  justify-content: center;
-  align-items: center;
   gap: 1rem;
+  align-items: center;
+  justify-content: center;
+
+  margin-bottom: 1.5rem;
 
   list-style: none;
-  margin-bottom: 1.5rem;
 `;
 
 export const CardContent = styled(Link)`
   display: flex;
-  justify-content: flex-start;
   flex-direction: column;
+  justify-content: flex-start;
 
-  border: 1px solid ${({ theme }) => theme.colors.secondary};
-  border-radius: 1rem;
-  padding: 0.5rem 1rem;
   width: 100%;
   max-width: 65rem;
+  padding: 0.5rem 1rem;
+
   background-color: ${({ theme }) => theme.colors.secondary};
-  box-shadow: 0px 0px 5px rgba(0, 0, 0, 0.5);
+  border: 1px solid ${({ theme }) => theme.colors.secondary};
+  border-radius: 1rem;
+  box-shadow: 0 0 5px rgb(0 0 0 / 50%);
+
   transition: 0.3s ease-in-out;
 
   &:hover {
