+++ src/app/notes/components/NotesContainer/index.styled.ts
@@ -1,14 +1,14 @@
 import styled from 'styled-components';
 
 export const Container = styled.div`
-  margin: 4rem auto;
-  max-width: 80rem;
-
   display: flex;
-  justify-content: flex-start;
   flex-direction: column;
-  align-items: center;
   gap: 2rem;
+  align-items: center;
+  justify-content: flex-start;
+
+  max-width: 80rem;
+  margin: 4rem auto;
 
   span {
     font-weight: bold;
@@ -18,9 +18,9 @@ export const Container = styled.div`
 
 export const TopContainer = styled.div`
   display: flex;
-  justify-content: space-between;
   flex-direction: column-reverse;
   align-items: center;
+  justify-content: space-between;
 
   width: 100%;
   max-width: 65rem;
@@ -31,49 +31,57 @@ export const TopContainer = styled.div`
 `;
 
 export const Wave = styled.span`
-  @keyframes waveAnimation {
+  @keyframes wave-animation {
     0% {
       transform: rotate(0deg);
     }
+
     10% {
       transform: rotate(14deg);
     }
+
     20% {
       transform: rotate(-8deg);
     }
+
     30% {
       transform: rotate(14deg);
     }
+
     40% {
       transform: rotate(-4deg);
     }
+
     50% {
       transform: rotate(10deg);
     }
+
     60% {
       transform: rotate(0deg);
     }
+
     100% {
       transform: rotate(0deg);
     }
   }
 
-  animation: waveAnimation 2.5s infinite;
   transform-origin: 70% 70%;
   display: inline-block;
+  animation: wave-animation 2.5s infinite;
 `;
 
 export const Number = styled.span`
-  font-weight: bold;
   font-size: ${({ theme }) => theme.typography.fontSize.h3}rem;
+  font-weight: bold;
   color: ${({ theme }) => theme.colors.primary};
 `;
 
 export const List = styled.ul`
+  overflow-y: scroll;
+
   width: 100%;
   max-width: 55rem;
   max-height: 50vh;
-  overflow-y: scroll;
   padding: 1rem;
 
   border: 1px transparent solid;
