+++ src/components/ui/Button/index.styled.tsx
@@ -1,16 +1,21 @@
 import styled from 'styled-components';
 
 const Button = styled.button`
-  border: none;
-  color: ${({ theme }) => theme.colors.secondary};
+  cursor: pointer;
+
+  display: inline-block;
+
+  padding: 0.5rem 1rem;
+
   font-family: ${({ theme }) => theme.typography.fontFamily.ultra};
   font-size: ${({ theme }) => theme.typography.fontSize.h5}rem;
+  color: ${({ theme }) => theme.colors.secondary};
+
   background-color: ${({ theme }) => theme.colors.primary};
-  padding: 0.5rem 1rem;
+  border: none;
   border-radius: 1rem;
+
   transition: 0.4s ease-in-out;
-  display: inline-block;
-  cursor: pointer;
 
   &:hover {
     color: ${({ theme }) => theme.colors.primary};
