+++ src/components/ui/FormikUi/FormikSelectField/index.styled.ts
@@ -1,8 +1,8 @@
 import styled from 'styled-components';
 
 export const CustomFormikField = styled.select`
+  cursor: pointer;
+  padding: 0.5rem;
   border: 1px solid ${({ theme }) => theme.colors.grey};
   border-radius: 1rem;
-  padding: 0.5rem;
-  cursor: pointer;
 `;
