+++ src/components/ui/FormikUi/FormikTextField/index.styled.ts
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 
 export const CustomFormikField = styled.input`
+  padding: 0.5rem;
   border: 1px solid ${({ theme }) => theme.colors.grey};
   border-radius: 1rem;
-  padding: 0.5rem;
 `;
