+++ src/components/ui/FormikUi/common.styled.ts
@@ -5,13 +5,13 @@ export const FormikContainer = styled.div`
   flex-direction: column;
   gap: 0.2rem;
 
-  margin: 0 auto;
-  margin-bottom: 1rem;
   width: 100%;
   max-width: 50rem;
+  margin: 0 auto;
+  margin-bottom: 1rem;
 `;
 
 export const FormikErrorText = styled.p`
-  color: ${({ theme }) => theme.colors.error};
   margin-top: 0.2rem;
+  color: ${({ theme }) => theme.colors.error};
 `;
