+++ src/components/ui/index.ts
@@ -1,7 +1,7 @@
 // General UI
-export { default as Button } from './Button';
-export { default as ButtonLink } from './ButtonLink';
-export { default as Container } from './Container';
+export { default as Button } from './Button/index.styled';
+export { default as ButtonLink } from './ButtonLink/index.styled';
+export { default as Container } from './Container/index.styled';
 export { default as Typography } from './Typography';
 
 // Formik UI
